+++ meta_package_manager/base.py
@@ -88,6 +88,9 @@ class PackageManager(object):
     def __init__(self):
         # Tell the manager either to raise or continue on errors.
         self.raise_on_cli_error = False
+        # Some managers have the ability to report or ignore packages
+        # possessing their own auto-update mecanism.
+        self.ignore_auto_updates = True
         # Log of all encountered CLI errors.
         self.cli_errors = []